+++ src/vitest.ts
@@ -1,7 +1,13 @@
 import { expect } from 'vitest';
 
 import { matchers } from './matchers.js';
-import type { CookieExpectation, HeaderExpectation } from './types.js';
+import type {
+  CookieExpectation,
+  FormDataExpectation,
+  HeaderExpectation,
+  QueryExpectation,
+  UrlExpectation,
+} from './types.js';
 
 expect.extend(matchers);
 
@@ -12,7 +18,11 @@ declare module 'vitest' {
     toHaveStatus(expected: number): T;
     toHaveHeader(name: string, expected?: HeaderExpectation): T;
     toHaveJson(expected: unknown): Promise<T>;
+    toHaveFormData(expected: FormDataExpectation): Promise<T>;
+    toHaveMethod(expected: string): T;
+    toHaveQuery(expected: QueryExpectation): T;
     toHaveText(expected: string | RegExp): Promise<T>;
+    toHaveUrl(expected: UrlExpectation): T;
     toRedirectTo(location: string | URL, status?: number): T;
     toSetCookie(name: string, expected?: CookieExpectation): T;
   }
@@ -20,6 +30,9 @@ declare module 'vitest' {
   interface AsymmetricMatchersContaining {
     toHaveStatus(expected: number): unknown;
     toHaveHeader(name: string, expected?: HeaderExpectation): unknown;
+    toHaveMethod(expected: string): unknown;
+    toHaveQuery(expected: QueryExpectation): unknown;
+    toHaveUrl(expected: UrlExpectation): unknown;
     toRedirectTo(location: string | URL, status?: number): unknown;
     toSetCookie(name: string, expected?: CookieExpectation): unknown;
   }